% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{add_rownames}
\alias{add_rownames}
\alias{use_as_rownames}
\title{Converting between rownames and a column}
\usage{
add_rownames(df, var = "rowname")

use_as_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\description{
add_rownames: Convert row names to an explicit variable.

use_as_rownames: Inverted operation of add_rownames, returning a data.frame
}
\examples{
tbl_df(mtcars)

mtcars_tbl <- add_rownames(mtcars)

mtcars_data.frame <- use_as_rownames(mtcars_tbl)
}

