% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as.tibble}
\title{Deprecated coercion functions}
\usage{
as_data_frame(x, ...)

as.tibble(x, ...)
}
\description{
Use \code{\link[=as_tibble]{as_tibble()}} instead, but mind the new signature and semantics.
}
\keyword{internal}
