% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{col2rowname}
\alias{col2rowname}
\title{Convert a column of tbl into rownames of data.frame.}
\usage{
col2rowname(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\description{
Inverted operation of add_rownames
}
\examples{
tbl_df(mtcars)

mtcars_tbl <- add_rownames(mtcars)
mtcars_data.frame <- col2rowname(mtcars_tbl)
}
\seealso{
\code{\link{add_rownames}}
}

