% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decor.R
\name{decor}
\alias{decor}
\alias{decor_num}
\alias{decor_chr}
\title{Decorate a column}
\usage{
decor_num(x, sigfig = getOption("pillar.sigfig", 3))

decor_chr(x, min_width = 3L)
}
\arguments{
\item{x}{The column data.}

\item{min_width}{Avoid shortening the column below this
width.}

\item{signif}{The number of significant digits.}
}
\description{
If the default data output in a tibble doesn't work for you,
the \code{decor_()} family of functions allows overriding it
on a column-by-column basis. Call \code{guess_decor()} for a
tibble to guess decorations for all (undecorated)
columns in a tibble.
}
\details{
Use the scoped functions in \pkg{dplyr} to apply decorations
to multiple columns.
}
\section{Lifecycle}{

These functions are experimental.
}

