% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{data_frame}
\alias{tibble_}
\alias{data_frame_}
\alias{lst_}
\alias{as_data_frame}
\alias{as.tibble}
\alias{frame_data}
\title{Deprecated functions}
\usage{
data_frame(...)

tibble_(xs)

data_frame_(xs)

lst_(xs)

as_data_frame(x, ...)

as.tibble(x, ...)

frame_data(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=tibble]{tibble()}} instead of \code{data_frame()}.

Use \link[rlang:quasiquotation]{quasiquotation} instead of \code{tibble_()}, \code{data_frame_()}, and \code{lst_()}.

Use \code{\link[=as_tibble]{as_tibble()}} instead of \code{as_data_frame()} or \code{as.tibble()},
but mind the new signature and semantics.

Use \code{\link[=tribble]{tribble()}} instead of \code{frame_data()}.
}
\keyword{internal}
