% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dput-tribble.R
\name{dput_tribble}
\alias{dput_tribble}
\title{Write a tibble or data.frame to a file or the screen as a call to tribble}
\usage{
dput_tribble(x, file = "")
}
\arguments{
\item{x}{a \code{tibble} or other \code{data.frame}}

\item{file}{either a character string naming a file or a connection.
"" indicates output to the console}
}
\value{
the first argument invisibly.
}
\description{
\code{tribble()} creates \link{tibble}s using an easier to read row-by-row layout.
\code{dput_tribble()} writes to the screen or a file the appropriate call to
\code{tribble()} to recreate an existing \code{tibble} or \code{data.frame}.
Please see \link{tibble-package} for a general introduction.
}
\details{
\code{dput_tribble} currently handles character, factor, integer, numeric,
logical and list column types.
}
\examples{
dput_tribble(iris)
}
