% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{is.data_frame}
\alias{is.data_frame}
\alias{is_data_frame}
\title{Test if the object is a tibble.}
\usage{
is.data_frame(x)

is_data_frame(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tbl_df} class.
}
\description{
Test if the object is a tibble.
}

