% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{is.tibble}
\alias{is.tibble}
\title{Deprecated test for tibble-ness}
\usage{
is.tibble(x)
}
\arguments{
\item{x}{An object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}

Please use \code{\link[=is_tibble]{is_tibble()}} instead.
}
\keyword{internal}
