% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{is_tibble}
\alias{is_tibble}
\title{Test if the object is a tibble}
\usage{
is_tibble(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tbl_df} class.
}
\description{
\Sexpr[results=rd, stage=render]{tibble:::lifecycle("stable")}

This function returns \code{TRUE} for tibbles or subclasses thereof,
and \code{FALSE} for all other objects, including regular data frames.
}
