% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.r
\name{obj_sum}
\alias{obj_sum}
\alias{type_sum}
\alias{tbl_sum}
\alias{is_vector_s3}
\title{Provide a succinct summary of an object}
\usage{
obj_sum(x)

type_sum(x)

tbl_sum(x)

is_vector_s3(x)
}
\arguments{
\item{x}{an object to summarise. Generally only methods of atomic vectors
and variants have been implemented.}
}
\description{
\code{type_sum()} gives a brief summary of object type. Objects that commonly
occur in a data frame should return a string with four or less characters.

\code{obj_sum()} also includes the size of the object if \code{is_vector_s3()}
is \code{TRUE}.

\code{tbl_sum()} gives a brief textual description of a table-like object,
which should include the dimensions, the data source, and possible grouping
(for \code{dplyr}).  The default implementation forwards to \code{obj_sum()}.
}
\examples{
obj_sum(1:10)
obj_sum(matrix(1:10))
obj_sum(Sys.Date())
obj_sum(Sys.time())
obj_sum(mean)
}
\keyword{internal}
