% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-format.r
\name{print.tbl_df}
\alias{print.tbl_df}
\title{Legacy help page for compatibility with existing packages}
\description{
\Sexpr[results=rd, stage=render]{tibble:::lifecycle("archived")}

Please see \code{\link[=print.tbl]{print.tbl()}} for the print method for tibbles.
}
\keyword{internal}
