% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-column-names.R
\name{repair_names}
\alias{repair_names}
\title{Check validity/format of a data.frame or tbl.}
\usage{
repair_names(x, prefix = "V", sep = "")
}
\arguments{
\item{x}{named object}

\item{prefix}{character, the prefix to use for new column names}

\item{sep}{character, a character string to separate the terms. Not
\code{NA_character_}.}
}
\value{
input with reparied names
}
\description{
Ensure the object (e.g., tbl, data.frame or list) has legitimate and unique
names without leading or trailing blanks.
}
\details{
The function does not change any names that are correct already.
}
\examples{
repair_names(list(3, 4, 5)) # works for lists, too
repair_names(mtcars) # a no-op
tbl <- as_data_frame(structure(list(3, 4, 5), class = "data.frame"))
repair_names(tbl)
}

