% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df.R
\name{tbl_rbind}
\alias{tbl_rbind}
\alias{tbl_cbind}
\title{Binding rows and columns}
\usage{
tbl_rbind(...)

tbl_cbind(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

In tibble 4.0.0, the \code{\link[=rbind]{rbind()}} and \code{\link[=cbind]{cbind()}} methods
will become stricter.
These functions are designed to help the transition to those stricter
versions.
Replace calls to \code{rbind()} and \code{cbind()} with \code{tbl_rbind()} and \code{tbl_cbind()},
respectively, to preview how your code will behave after the upgrade.
Set the \code{tibble.rbind_trace} or \code{tibble.cbind_trace} options to \code{TRUE}
to enable diagnostic information.
}
