% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.r
\name{tbl_sum}
\alias{tbl_sum}
\title{Provide a succinct summary of an object}
\usage{
tbl_sum(x)
}
\description{
\code{tbl_sum()} gives a brief textual description of a table-like object,
which should include the dimensions, the data source, and possible grouping
(for \code{dplyr}).  The default implementation forwards to \code{\link[colformat:obj_sum]{colformat::obj_sum()}}.
}
\seealso{
\code{\link[colformat:type_sum]{colformat::type_sum()}}, \code{\link[colformat:is_vector_s3]{colformat::is_vector_s3()}}
}
