% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{trunc_mat}
\alias{trunc_mat}
\title{Legacy printing}
\usage{
trunc_mat(x, n = NULL, width = NULL, n_extra = NULL)
}
\arguments{
\item{x}{Object to format or print.}

\item{n}{Number of rows to show.
If \code{NULL}, the default, will print all rows if less than option \code{tibble.print_max}.
Otherwise, will print \code{tibble.print_min} rows.}

\item{width}{Width of text output to generate.
This defaults to \code{NULL}, which means use \code{getOption("tibble.width")} or (if also \code{NULL}) \code{getOption("width")};
the latter displays only the columns that fit on one screen.
You can also set \code{options(tibble.width = Inf)} to override this default and always print all columns,
this may be slow for very wide tibbles.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble.
If \code{NULL}, the default, will print information about at most \code{tibble.max_extra_cols} extra columns.}
}
\value{
An object with a \code{print()} method that will print the input similarly to a tibble.
The internal data format is an implementation detail, do not rely on it.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
As of tibble 3.1.0, printing is handled entirely by the \pkg{pillar} package.
Do not use this function.
If you implement a package that extend tibble,
the printed output can be customized in various ways.
See \code{vignette("extending", package = "pillar")} for details.
}
\keyword{internal}
