% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble.R
\name{turn_frame_data_into_frame_matrix}
\alias{turn_frame_data_into_frame_matrix}
\title{Turn frame data into a frame_matrix}
\usage{
turn_frame_data_into_frame_matrix(data)
}
\arguments{
\item{data}{to be a list with 2 elements,
\code{frame_names}-a character vector, and \code{frame_rest}-a list}
}
\description{
Turn frame data into a frame_matrix
}

