% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble.R
\name{turn_frame_data_into_tibble}
\alias{turn_frame_data_into_tibble}
\title{Turn frame data into a tibble}
\usage{
turn_frame_data_into_tibble(data)
}
\arguments{
\item{data}{a list with 2 elements,
\code{frame_names}-a character vector, and \code{frame_rest}-a list}
}
\description{
Turn frame data into a tibble
}

