% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{use_as_rownames}
\alias{use_as_rownames}
\title{Convert a column of tbl into rownames of data.frame.}
\usage{
use_as_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\description{
Inverted operation of add_rownames
}
\examples{
tbl_df(mtcars)

mtcars_tbl <- add_rownames(mtcars)
mtcars_data.frame <- use_as_rownames(mtcars_tbl)
}
\seealso{
\code{\link{add_rownames}}
}

