% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View an object}
\usage{
view(x, title = NULL, ..., n = NULL)
}
\arguments{
\item{x}{The object to display.}

\item{title}{The title to use for the display, by default the deparsed expression is used.}

\item{...}{Unused, must be empty.}

\item{n}{Maximum number of rows to display.
Only used if \code{x} is not a data frame.
Uses the \code{view_max} \link[=tibble_options]{option} by default.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calls \code{\link[utils:View]{utils::View()}} on the input and returns it, invisibly.
If the input is not a data frame, it is processed using a variant of \code{as.data.frame(head(x, n))}.
A message is printed if the number of rows exceeds \code{n}.
This function has no effect in non\link{interactive} sessions.
}
\details{
The RStudio IDE overrides \code{utils::View()}, this is picked up correctly.
}
