% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_row}
\alias{add_row}
\title{Add rows to a data frame}
\usage{
add_row(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{Data frame to append to.}

\item{...}{Name-value pairs. If you don't supply the name of a variable,
it'll be given the value \code{NA}.}

\item{.before, .after}{One-based row index where to add the new rows,
default: after last row}
}
\description{
This is a convenient way to add one or more rows of data to an existing data
frame. See \code{\link{tribble}} for an easy way to create an complete
data frame row-by-row.
}
\examples{
# add_row ---------------------------------
df <- tibble(x = 1:3, y = 3:1)

add_row(df, x = 4, y = 0)

# You can specify where to add the new rows
add_row(df, x = 4, y = 0, .before = 2)

# You can supply vectors, to add multiple rows (this isn't
# recommended because it's a bit hard to read)
add_row(df, x = 4:5, y = 0:-1)

# Absent variables get missing values
add_row(df, x = 4)

# You can't create new variables
\dontrun{
add_row(df, z = 10)
}
}
\seealso{
Other addition: \code{\link{add_column}}
}

