% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.data.frame}
\alias{as_tibble.list}
\alias{as_tibble.matrix}
\alias{as_tibble.table}
\alias{as_tibble.NULL}
\alias{as_tibble.default}
\title{Coerce vectors, lists, and matrices to data frames}
\usage{
as_tibble(x, ..., .rows = NULL, .name_repair = c("check_unique",
  "unique", "syntactic", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL))

\method{as_tibble}{data.frame}(x, validate = TRUE, ..., .rows = NULL,
  .name_repair = c("check_unique", "unique", "syntactic", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL))

\method{as_tibble}{list}(x, validate = TRUE, ..., .rows = NULL,
  .name_repair = c("check_unique", "unique", "syntactic", "minimal"))

\method{as_tibble}{matrix}(x, ...)

\method{as_tibble}{table}(x, `_n` = "n", ..., n = `_n`)

\method{as_tibble}{NULL}(x, ...)

\method{as_tibble}{default}(x, ...)
}
\arguments{
\item{x}{A vector, list, or matrix. If a list, each element must have the same length.}

\item{...}{Other arguments passed on to individual methods.}

\item{.rows}{The number of rows, useful to create a 0-column tibble or
just as an additional check.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"syntactic"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[=as_function]{as_function()}}
}

See \link{name-repair} for more details on these terms and the strategies used
to enforce them.}

\item{rownames}{Treatment of existing row names (for data frames and matrices):
\itemize{
\item \code{NULL}: remove row names
\item \code{NA}: keep row names (default, for compatibility; will issue a warning if row names
are present in the input).
\item A string: the name of the new column that will contain the existing row names,
which are no longer present in the result.
}}

\item{_n, validate}{For compatibility only, do not use for new code.}

\item{n}{Name for count column, default: \code{"n"}.}
}
\description{
This is an S3 generic. Methods are included for:
\itemize{
\item data frames (adds \code{"tbl_df"} classes)
\item tibbles (returns unchanged input)
\item lists and vectors (treats each element as a column)
\item matrices
\item tables
\item default (coerces to a list)
}
}
\details{
Vectors and unknown input is coerced to a list (unlike \code{as.data.frame()}):
for vectors, the returned tibble has as many columns as the vector has
elements.

\code{\link[=as.data.frame]{as.data.frame()}} is effectively a thin wrapper around \code{data.frame},
and hence is rather slow (because it calls \code{\link[=data.frame]{data.frame()}} on each element
before \link{cbind}ing together). \code{as_tibble} is a new S3 generic
with more efficient methods for matrices and data frames.

The default behavior is to silently remove row names.
New code should explicitly convert row names to a new column using the \code{rownames}
argument.
For existing code that relies on row names remaining, call
\code{pkgconfig::set_config("tibble::rownames" = NA)} in your script or in your package's
\code{\link[=.onLoad]{.onLoad()}}  function.
}
\examples{
l <- list(x = 1:500, y = runif(500), z = 500:1)
df <- as_tibble(l)

m <- matrix(rnorm(50), ncol = 5)
colnames(m) <- c("a", "b", "c", "d", "e")
df <- as_tibble(m)

as_tibble(1:3, .name_repair = "unique")

# For list-like inputs, as_tibble() is considerably simpler than as.data.frame()
\dontrun{
if (requireNamespace("bench", quietly = TRUE)) {
  l2 <- replicate(26, sample(letters), simplify = FALSE)
  names(l2) <- letters
  bench::mark(
    as_tibble(l2, .name_repair = "syntactic"),
    as_tibble(l2, .name_repair = "unique"),
    as_tibble(l2, .name_repair = "minimal"),
    as_tibble(l2),
    as.data.frame(l2),
    check = FALSE
  )
}
}
}
\seealso{
\code{\link[=enframe]{enframe()}} converts a vector to a data frame with values in rows,
\link{name-repair} documents the details of name repair.

\code{\link[pkgconfig:set_config]{pkgconfig::set_config()}}
}
