% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{as_data_frame}
\alias{as.tibble}
\alias{frame_data}
\title{Deprecated functions}
\usage{
as_data_frame(x, ...)

as.tibble(x, ...)

frame_data(...)
}
\description{
Use \code{\link[=as_tibble]{as_tibble()}} instead of \code{as_data_frame()} or \code{as.tibble()}, but mind the new signature and semantics.

Use \code{\link[=tribble]{tribble()}} instead of \code{frame_data()}.
}
\keyword{internal}
