% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R
\name{enframe}
\alias{enframe}
\alias{deframe}
\title{Converting vectors to data frames, and vice versa}
\usage{
enframe(x, name = "name", value = "value")

deframe(x)
}
\arguments{
\item{x}{For \code{enframe()}, a vector.
For \code{deframe()}, a data frame with two columns.
Passing a one-column data frame is deprecated, use \code{x[[1]]} instead.}

\item{name, value}{Names of the columns that store the names and values.
Passing \code{NULL} is deprecated, use \code{\link[=as_tibble_col]{as_tibble_col()}} instead.}
}
\value{
For \code{enframe()}, a \link{tibble} with two columns.
For \code{deframe()}, a named vector.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

\code{enframe()} converts named atomic vectors or lists to two-column data frames.
For a list, the result will be a nested tibble with a column of type \code{list}.
For unnamed vectors, the natural sequence is used as name column.

\code{deframe()} converts two-column data frames to a named vector or list,
using the first column as name and the second column as value.
}
\examples{
enframe(1:3)
enframe(c(a = 5, b = 7))
enframe(list(one = 1, two = 2:3, three = 4:6))
enframe(trees)
enframe(matrix(1:6, ncol = 2))
deframe(enframe(3:1))
deframe(tibble(a = 1:3))
deframe(tibble(a = as.list(1:3)))
}
