% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{is.tibble}
\alias{is.tibble}
\title{Deprecated test for tibble-ness}
\usage{
is.tibble(x)
}
\arguments{
\item{x}{An object}
}
\description{
\Sexpr[results=rd, stage=render]{tibble:::lifecycle("soft-deprecated")}

Please use \code{\link[=is_tibble]{is_tibble()}} instead.
}
\keyword{internal}
