% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{is.tibble}
\alias{is.tibble}
\alias{is_tibble}
\title{Test if the object is a tibble.}
\usage{
is.tibble(x)

is_tibble(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tbl_df} class.
}
\description{
Test if the object is a tibble.
}
