% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{is_tibble}
\alias{is_tibble}
\alias{is.tibble}
\title{Test if the object is a tibble}
\usage{
is_tibble(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tbl_df} class.
}
\description{
This function returns \code{FALSE} for regular data frames and \code{TRUE} for tibbles.
}
