% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst.R
\name{lst}
\alias{lst}
\alias{lst_}
\title{Build a list}
\usage{
lst(...)

lst_(xs)
}
\description{
\code{lst()} is similar to \code{\link[=list]{list()}}, but like \code{tibble()}, it
evaluates its arguments lazily and in order, and automatically adds names.
}
