% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_tibble}
\alias{new_tibble}
\alias{new_valid_tibble}
\alias{set_tibble_class}
\title{Constructor}
\usage{
new_tibble(x, ..., names = NULL, nrow = NULL, subclass = NULL)
}
\arguments{
\item{x}{A tibble-like object}

\item{...}{Passed on to \code{\link[=structure]{structure()}}}

\item{names}{If non-\code{NULL}, override the names of \code{x}}

\item{nrow}{The number of rows, required}

\item{subclass}{Subclasses to assign to the new object, default: none}
}
\description{
Creates a subclass of a tibble.
This function is mostly useful for package authors that implement subclasses
of a tibble, like \pkg{sf} or \pkg{tibbletime}.
}
\details{
\code{x} must be a list.

The \code{new_tibble()} constructor requires an \code{nrow} argument.
It makes sure that the \code{row.names} attribute
is consistent with the data before returning.

If provided, the \code{names} argument is used to override the names of \code{x}.

The \code{...} argument allows adding more attributes to the subclass.

\code{x} must have names (or be empty),
but the names are not checked for correctness.

1d arrays are always converted to vectors.

The \code{row.names} attribute will be created from the \code{nrow} argument,
overriding any existing attribute of this name in \code{x} or in the \code{...}
arguments.

The \code{class} attribute of the returned object always consists of
\code{c("tbl_df", "tbl", "data.frame")}. If the \code{subclass} argument is set,
it will be prepended to that list of classes.
}
\examples{
# The nrow argument is always required:
new_tibble(list(a = 1:3, b = 4:6), nrow = 3)

# Existing row.names attributes are ignored:
try(new_tibble(iris, nrow = 3))

# The length of all columns must be consistent with the nrow argument:
try(new_tibble(list(a = 1:3, b = 4:6), nrow = 2))
}
