% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tbl_df}
\alias{print.tbl_df}
\title{Legacy help page for compatibility with existing packages}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#superseded'><img src='figures/lifecycle-superseded.svg' alt='Superseded lifecycle'></a>}}{\strong{Superseded}}

Please see \code{\link[=print.tbl]{print.tbl()}} for the print method for tibbles.
}
\keyword{internal}
