% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{obj_sum}
\alias{reexports}
\alias{type_sum}
\alias{reexports}
\alias{is_vector_s3}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{colformat}{\code{\link[colformat]{obj_sum}}, \code{\link[colformat]{type_sum}}, \code{\link[colformat]{is_vector_s3}}}
}}

