% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-name.R, R/pipe.R, R/tbl_sum.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{has_name}
\alias{\%>\%}
\alias{obj_sum}
\alias{type_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{pillar}{\code{\link[pillar:type_sum]{obj_sum}}, \code{\link[pillar]{type_sum}}}

  \item{rlang}{\code{\link[rlang]{has_name}}}
}}

