% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-name.R, R/type-sum.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{has_name}
\alias{obj_sum}
\alias{type_sum}
\alias{is_vector_s3}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pillar}{\code{\link[pillar]{obj_sum}}, \code{\link[pillar]{type_sum}}, \code{\link[pillar]{is_vector_s3}}}

  \item{rlang}{\code{\link[rlang]{has_name}}}
}}

