% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse.R, R/has-name.R, R/pipe.R, R/tbl_sum.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glimpse}
\alias{has_name}
\alias{\%>\%}
\alias{tbl_sum}
\alias{obj_sum}
\alias{type_sum}
\alias{size_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{pillar}{\code{\link[pillar]{glimpse}}, \code{\link[pillar:type_sum]{obj_sum}}, \code{\link[pillar:type_sum]{size_sum}}, \code{\link[pillar]{tbl_sum}}, \code{\link[pillar]{type_sum}}}

  \item{rlang}{\code{\link[rlang]{has_name}}}
}}

