% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-column-names.R
\name{repair_names}
\alias{repair_names}
\title{Check validity/format of a data.frame or tbl.}
\usage{
repair_names(x, prefix = "V", sep = "")
}
\arguments{
\item{x}{tbl, tbl_df, or data.frame}

\item{prefix}{character, the prefix to use for new column names}

\item{sep}{character, a character string to separate the terms. Not
\code{NA_character_}.}
}
\value{
data.frame with possibly-repaired column names
}
\description{
Ensure the tbl or data.frame has legitimate and optionally unique
column names.
}

