% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R, R/dataframe.R
\name{has_rownames}
\alias{column_to_rownames}
\alias{has_rownames}
\alias{remove_rownames}
\alias{rownames_to_column}
\title{Row names}
\usage{
has_rownames(df)

remove_rownames(df)

rownames_to_column(df, var = "rowname")

column_to_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame}

\item{var}{Name of variable to use}
}
\description{
\code{has_rownames} checks if a data frame has row names.

\code{remove_rownames} removes all row names.

\code{rownames_to_column} convert row names to an explicit variable.

\code{column_to_rownames} convert a column variable to row names. This is an
inverted operation of \code{rownames_to_column}.
}
\examples{
has_rownames(mtcars)
has_rownames(iris)
rownames(remove_rownames(mtcars))

rownames_to_column(mtcars)

mtcars_tbl <- rownames_to_column(tbl_df(mtcars))
mtcars_tbl

column_to_rownames(mtcars_tbl)
}

