% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{rownames_to_column}
\alias{column_to_rownames}
\alias{rownames_to_column}
\title{Conversion between rownames and a column in data frame}
\usage{
rownames_to_column(df, var = "rowname")

column_to_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\description{
\code{rownames_to_column} convert row names to an explicit variable.

\code{column_to_rownames} convert a column variable to row names. This is an
inverted operation of \code{rownames_to_column}.
}
\examples{
rownames_to_column(mtcars)

mtcars_tbl <- rownames_to_column(tbl_df(mtcars))
mtcars_tbl

column_to_rownames(mtcars_tbl)
}

