% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{add_rownames}
\alias{add_rownames}
\alias{use_as_rownames}
\title{Conversion between rownames and a column in data frame}
\usage{
add_rownames(df, var = "rowname")

use_as_rownames(df, var = "rowname")
}
\arguments{
\item{df}{Input data frame with rownames.}

\item{var}{Name of variable to use}
}
\value{
add_rownames: c("tbl_df", "tbl", "data.frame")

use_as_rownames: data.frame
}
\description{
\code{add_rownames} Convert row names to an explicit variable.

\code{use_as_rownames} Convert a column variable to row names. This is an
inverted operation of add_rownames, returning a data.frame
}
\examples{
tbl_df(mtcars)

(mtcars_tbl <- add_rownames(mtcars) )

head ( use_as_rownames(mtcars_tbl) )
df <- mtcars_tbl
var = "rowname"
}

