% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{subsetting}
\alias{subsetting}
\alias{[[.tbl_df}
\alias{$.tbl_df}
\alias{[.tbl_df}
\title{Subsetting tibbles}
\usage{
\method{[[}{tbl_df}(x, i, j, ..., exact = TRUE)

\method{$}{tbl_df}(x, name)

\method{[}{tbl_df}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{data frame.}

\item{i, j}{Row and column indices. If \code{j} is omitted, \code{i} is used as column index.}

\item{...}{Ignored.}

\item{exact}{Ignored, with a warning.}

\item{name}{
    A literal character string or a \link{name} (possibly \link{backtick}
    quoted).}

\item{drop}{Coerce to a vector if fetching one column via \code{tbl[, j]} .
Default \code{FALSE}, ignored when accessing a column via \code{tbl[j]} .}
}
\description{
Accessing columns, rows, or cells via \code{$}, \code{[[}, or \code{[} works very much like
\link[base:Extract.data.frame]{with regular data frames}. However, the
behavior is different for tibbles and data frames in some cases:
\itemize{
\item \code{[} always returns a tibble by default, even if
only one column is accessed.  The \code{drop} argument is supported, but you
have to request it explicitly.
\item Partial matching of column names with \code{$} and \code{[[} is not supported, a
warning is given and \code{NULL} is returned.
}

Unstable return type and implicit partial matching can lead to surprises and
bugs that are hard to catch. If you rely on code that requires the original
data frame behavior, coerce to a data frame via \code{\link[=as.data.frame]{as.data.frame()}}.
}
\examples{
df <- data.frame(a = 1:3, bc = 4:6)
tbl <- tibble(a = 1:3, bc = 4:6)

# Same behavior:
tbl[1, ]
tbl[1, c("bc", "a")]
tbl[, c("bc", "a")]
tbl[c("bc", "a")]
tbl["a"]
tbl$a
tbl[["a"]]

# Different behavior:
df[, "a"]
tbl[, "a"]
tbl[, "a", drop = TRUE]
as.data.frame(tbl)[, "a"]

df[1, , drop = TRUE]
tbl[1, , drop = TRUE]
as.list(tbl[1, ])

df$b
tbl$b

df[["b", exact = FALSE]]
tbl[["b", exact = FALSE]]

df$bd <- c("n", "e", "w")
tbl$bd <- c("n", "e", "w")
df$b
tbl$b

df$b <- 7:9
tbl$b <- 7:9
df$b
tbl$b
}
