% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-df.r
\name{tbl_df-class}
\alias{tbl_df-class}
\alias{tbl_df}
\title{\code{tbl_df} class}
\description{
Lorem ipsum
}
\section{tibbles vs. data frames}{


Text is currently just taken from old description
\itemize{
\item Never coerces inputs (i.e. strings stay as strings!).
\item Never adds \code{row.names}.
\item Only recycles length 1 inputs.
\item Automatically adds column names.
\item Doesn't munge column names.
\item Evaluates its arguments lazily and in order.
\item Adds \code{tbl_df} class to output.
}
}

\section{Methods}{


\code{tbl_df} implements four important base methods:

\describe{
\item{print}{By default only prints the first 10 rows (at most 20), and the
columns that fit on screen; see \code{\link[=print.tbl]{print.tbl()}}}
\item{\code{[}}{Does not simplify (drop) by default, returns a data frame}
\item{\code{[[}, \code{$}}{Calls \code{\link[=.subset2]{.subset2()}} directly,
so is considerably faster. Returns \code{NULL} if column does not exist,
\code{$} warns.}
}
}

\section{Important functions}{

\code{\link[=tibble]{tibble()}} and \code{\link[=tribble]{tribble()}} for construction,
\code{\link[=as_tibble]{as_tibble()}} for coercion,
and \code{\link[=print.tbl]{print.tbl()}} and \code{\link[=glimpse]{glimpse()}} for display.
}

\seealso{
\code{\link[=tibble]{tibble()}}, \code{\link[=as_tribble]{as_tribble()}}, \code{\link[=tribble]{tribble()}}
}
