% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_sum.R
\name{tbl_sum}
\alias{tbl_sum}
\title{Provide a succinct summary of an object}
\usage{
tbl_sum(x, ...)
}
\arguments{
\item{x}{Object to summarise.}
}
\value{
A named character vector, describing the dimensions in the first element
and the data source in the name of the first element.
}
\description{
\code{tbl_sum()} gives a brief textual description of a table-like object,
which should include the dimensions and the data source in the first element,
and additional information in the other elements (such as grouping for \pkg{dplyr}).
The default implementation forwards to \code{\link[pillar:type_sum]{pillar::obj_sum()}}.
}
\details{
This generic will be moved to \pkg{pillar}, and reexported from there
as soon as it becomes available.
}
\seealso{
\code{\link[pillar:type_sum]{pillar::type_sum()}}
}
