% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble-package.R
\docType{package}
\name{tibble-package}
\alias{tibble-package}
\title{tibble: Simple Data Frames}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides a 'tbl_df' class (the 'tibble') that provides
    stricter checking and better formatting than the traditional data frame.
}
\details{
The tibble package provides utilities for handling \strong{tibbles}, where
"tibble" is a colloquial term for the S3 \code{\link{tbl_df}} class. The \code{\link{tbl_df}}
class is a special case of the base \code{\link[base:data.frame]{data.frame}}.
class, developed in response to lessons learned over many years of data
analysis with data frames.

General resources:
\itemize{
\item Website for the tibble package: \url{https://tibble.tidyverse.org}
\item \href{http://r4ds.had.co.nz/tibbles.html}{Tibbles chapter} in \emph{R for data
science}
}

Resources on specific topics:
\itemize{
\item Create a tibble: \code{\link[=tibble]{tibble()}}, \code{\link[=as_tibble]{as_tibble()}}, \code{\link[=tribble]{tribble()}}
\item Inspect a tibble: \code{\link[=print.tbl_df]{print.tbl_df()}}, \code{\link[=glimpse]{glimpse()}}
\item Details on the S3 \code{tbl_df} class: \code{\linkS4class{tbl_df}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://tibble.tidyverse.org/}
  \item \url{https://github.com/tidyverse/tibble}
  \item Report bugs at \url{https://github.com/tidyverse/tibble/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item Romain Francois \email{romain@r-enthusiasts.com} [contributor]
  \item RStudio [copyright holder]
}

}
