% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R, R/zzz.r
\docType{package}
\name{tibble-package}
\alias{tibble-package}
\title{Simple Data Frames}
\description{
Provides a 'tbl_df' class that offers better checking and
printing capabilities than traditional data frames.
}
\section{Package options}{

\describe{
\item{\code{tibble.print_max}}{Row number threshold: Maximum number of rows
  printed. Set to \code{Inf} to always print all rows.  Default: 20.}
\item{\code{tibble.print_min}}{Number of rows printed if row number
  threshold is exceeded. Default: 10.}
\item{\code{tibble.width}}{Output width. Default: \code{NULL} (use
  \code{width} option).}
}
}

