% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribble.R
\name{tribble}
\alias{tribble}
\alias{frame_data}
\title{Row-wise tibble creation}
\usage{
tribble(...)
}
\arguments{
\item{...}{Arguments specifying the structure of a \code{tibble}.
Variable names should be formulas, and may only appear before the data.
If a function needs to be applied to the data for a column (e.g. \code{factor}),
this should apear after the \code{~}, e.g. factor(column_A). Other arguments
cab be provided as needed.}
}
\value{
A \link{tibble}.
}
\description{
Create \link{tibble}s using an easier to read row-by-row layout.
This is useful for small tables of data where readability is
important.  Please see \link{tibble-package} for a general introduction.
}
\details{
\code{frame_data()} is an older name for \code{tribble()}. It will eventually
be phased out.
}
\examples{
tribble(
  ~colA, ~colB,
  "a",   1,
  "b",   2,
  "c",   3
)

# Conversion functions can be supplied, including additional
# paremeters if needed.
tribble(
  ~colA, ~factor(colB, levels = c("B", "A")),
  "X",   "A",
  "Y",   "B")

# tribble will create a list column if the value in any cell is
# not a scalar
tribble(
  ~x,  ~y,
  "a", 1:3,
  "b", 4:6
)
}
