% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-legacy.R
\name{trunc_mat}
\alias{trunc_mat}
\title{Prepare a tibble for formatting}
\usage{
trunc_mat(x, n = NULL, width = NULL, n_extra = NULL)
}
\arguments{
\item{x}{Object to format or print.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than option \code{tibble.print_max}. Otherwise, will print
\code{tibble.print_min} rows.}

\item{width}{Width of text output to generate. This defaults to \code{NULL}, which
means use \code{getOption("tibble.width")} or (if also \code{NULL})
\code{getOption("width")}; the latter displays only the columns that fit on one
screen. You can also set \code{options(tibble.width = Inf)} to override this
default and always print all columns, this may slow down printing
substantially.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the default,
will print information about at most \code{tibble.max_extra_cols} extra columns.}
}
\description{
\lifecycle{questioning}

For a tibble, returns an object that, when formatted with \code{\link[=format]{format()}},
produces the same output as formatting the tibble.
In other words, \code{format(trunc_mat(x, ...))} is the same as \code{format(x, ...)}.
}
\section{Life cycle}{

This function is in the questioning stage, see \link{custom-formatting}
for a modern experimental alternative.
}

