% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View your data}
\usage{
view(x, title = NULL, ...)
}
\description{
The default method of this generic calls \code{\link[=View]{View()}} (only in interactive
sessions) and returns the input.
}
\examples{
\dontrun{
view(iris)
}
}
