% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View an object}
\usage{
view(x, title = NULL, ...)
}
\arguments{
\item{x}{The object to display.}

\item{title}{The title to use for the display, by default
the deparsed expression is used.}

\item{...}{Unused, for extensibility.}
}
\description{
\Sexpr[results=rd, stage=render]{tibble:::lifecycle("experimental")}

Calls \code{\link[utils:View]{utils::View()}} on the input and returns it, invisibly.
The RStudio IDE overrides \code{utils::View()}, this is picked up
correctly.
}
